+++ library/Zend/Form/View/Helper/FormDateTimeSelect.php
@@ -23,7 +23,17 @@ class FormDateTimeSelect extends FormDateSelectHelper
      *
      * @var int
      */
-    protected $timeType = IntlDateFormatter::LONG;
+    protected $timeType;
+
+    /**
+     * @throws Exception\ExtensionsNotLoadedException if ext/intl is not present
+     */
+    public function __construct()
+    {
+        parent::__construct();
+
+        $this->timeType = IntlDateFormatter::LONG;
+    }
 
     /**
      * Invoke helper as function
