+++ library/Zend/I18n/Validator/DateTime.php
@@ -12,6 +12,7 @@ namespace Zend\I18n\Validator;
 use Locale;
 use IntlDateFormatter;
 use Traversable;
+use Zend\I18n\Exception as I18nException;
 use Zend\Validator\AbstractValidator;
 use Zend\Validator\Exception;
 
@@ -38,12 +39,12 @@ class DateTime extends AbstractValidator
     /**
      * @var int
      */
-    protected $dateType = IntlDateFormatter::NONE;
+    protected $dateType;
 
     /**
      * @var int
      */
-    protected $timeType = IntlDateFormatter::NONE;
+    protected $timeType;
 
     /**
      * Optional timezone
@@ -60,7 +61,7 @@ class DateTime extends AbstractValidator
     /**
      * @var int
      */
-    protected $calendar = IntlDateFormatter::GREGORIAN;
+    protected $calendar;
 
     /**
      * @var IntlDateFormatter
@@ -80,9 +81,21 @@ class DateTime extends AbstractValidator
      * Constructor for the Date validator
      *
      * @param array|Traversable $options
+     * @throws Exception\ExtensionNotLoadedException if ext/intl is not present
      */
     public function __construct($options = array())
     {
+        if (!extension_loaded('intl')) {
+            throw new I18nException\ExtensionNotLoadedException(sprintf(
+                '%s component requires the intl PHP extension',
+                __NAMESPACE__
+            ));
+        }
+
+        $this->dateType = IntlDateFormatter::NONE;
+        $this->timeType = IntlDateFormatter::NONE;
+        $this->calendar = IntlDateFormatter::GREGORIAN;
+
         parent::__construct($options);
 
         if (null === $this->locale) {
